Initial commit
Add the firmware code for the ESP32 microcontroller. This code is responsable for

reading the sensor MAX30100 data and then sends it to the REST API that will store

the data into the database to later consumption by other clients. This code also,

implements a WebServer for local visualization of the data that is being collected,

the user needs to be connected to the same network as the ESP32.